requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.6.0
    # via pydantic
asgiref==3.7.2
    # via django
beautifulsoup4==4.12.3
    # via bs4
bibtexparser==1.4.1
    # via pyzotero
boto3==1.34.64
    # via -r requirements.in
botocore==1.34.64
    # via
    #   boto3
    #   s3transfer
bs4==0.0.2
    # via -r requirements.in
build==1.1.1
    # via pip-tools
certifi==2024.2.2
    # via requests
chardet==5.2.0
    # via reportlab
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
cssselect2==0.7.0
    # via svglib
dj-database-url==2.1.0
    # via -r requirements.in
django==4.2.11
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-appconf
    #   django-classy-tags
    #   django-cms
    #   django-entangled
    #   django-filer
    #   django-formtools
    #   django-parler
    #   django-polymorphic
    #   django-sekizai
    #   django-select2
    #   django-storage-url
    #   django-storages
    #   django-treebeard
    #   djangocms-admin-style
    #   djangocms-alias
    #   djangocms-frontend
    #   djangocms-versioning
    #   easy-thumbnails
django-appconf==1.0.6
    # via django-select2
django-classy-tags==4.1.0
    # via
    #   -r requirements.in
    #   django-cms
    #   django-sekizai
django-cms==4.1.0
    # via
    #   -r requirements.in
    #   djangocms-alias
    #   djangocms-attributes-field
    #   djangocms-file
    #   djangocms-frontend
    #   djangocms-googlemap
    #   djangocms-history
    #   djangocms-icon
    #   djangocms-picture
    #   djangocms-snippet
    #   djangocms-style
    #   djangocms-text-ckeditor
    #   djangocms-transfer
    #   djangocms-versioning
    #   djangocms-video
django-entangled==0.5.4
    # via djangocms-frontend
django-filer==3.1.1
    # via
    #   djangocms-file
    #   djangocms-frontend
    #   djangocms-googlemap
    #   djangocms-picture
    #   djangocms-video
django-formtools==2.5.1
    # via django-cms
django-fsm==2.8.1
    # via djangocms-versioning
django-parler==2.3
    # via djangocms-alias
django-polymorphic==3.1.0
    # via django-filer
django-sekizai==4.1.0
    # via
    #   -r requirements.in
    #   django-cms
django-select2==8.1.2
    # via djangocms-frontend
django-storage-url==0.12.0
    # via -r requirements.in
django-storages==1.14.2
    # via django-storage-url
django-treebeard==4.7.1
    # via
    #   django-cms
    #   djangocms-googlemap
djangocms-admin-style==3.3.1
    # via
    #   -r requirements.in
    #   django-cms
djangocms-alias==2.0.0
    # via -r requirements.in
djangocms-attributes-field==3.0.0
    # via
    #   -r requirements.in
    #   djangocms-file
    #   djangocms-frontend
    #   djangocms-icon
    #   djangocms-picture
    #   djangocms-style
    #   djangocms-video
djangocms-file==3.0.1
    # via -r requirements.in
djangocms-frontend==1.2.2
    # via -r requirements.in
djangocms-googlemap==2.2.0
    # via -r requirements.in
djangocms-history==2.2.3
    # via -r requirements.in
djangocms-icon==2.1.0
    # via -r requirements.in
djangocms-picture==4.1.1
    # via -r requirements.in
djangocms-snippet==3.1.1
    # via -r requirements.in
djangocms-style==3.1.0
    # via -r requirements.in
djangocms-text-ckeditor==5.1.5
    # via
    #   -r requirements.in
    #   djangocms-frontend
djangocms-transfer==1.0.1
    # via -r requirements.in
djangocms-versioning==2.0.0
    # via -r requirements.in
djangocms-video==3.1.0
    # via -r requirements.in
dnspython==2.6.1
    # via email-validator
easy-thumbnails[svg]==2.8.5
    # via
    #   -r requirements.in
    #   django-filer
    #   djangocms-frontend
    #   djangocms-picture
email-validator==2.1.1
    # via hsmodels
feedparser==6.0.11
    # via pyzotero
furl==2.1.3
    # via django-storage-url
hs-restclient==1.3.7
    # via -r requirements.in
hsclient==1.0.0
    # via -r requirements.in
hsmodels==1.0.0
    # via hsclient
html5lib==1.1
    # via djangocms-text-ckeditor
idna==3.6
    # via
    #   email-validator
    #   requests
isodate==0.6.1
    # via rdflib
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==5.1.0
    # via svglib
oauthlib==3.2.2
    # via
    #   hs-restclient
    #   requests-oauthlib
orderedmultidict==1.0.1
    # via furl
packaging==24.0
    # via
    #   build
    #   django-cms
    #   djangocms-text-ckeditor
pillow==10.2.0
    # via
    #   djangocms-text-ckeditor
    #   easy-thumbnails
    #   reportlab
pip-tools==7.4.1
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
pydantic==2.6.4
    # via hsmodels
pydantic-core==2.16.3
    # via pydantic
pyparsing==3.1.2
    # via
    #   bibtexparser
    #   rdflib
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via botocore
pytz==2024.1
    # via
    #   -r requirements.in
    #   pyzotero
pyzotero==1.5.18
    # via -r requirements.in
rdflib==5.0.0
    # via hsmodels
reportlab==4.1.0
    # via
    #   easy-thumbnails
    #   svglib
requests==2.31.0
    # via
    #   hs-restclient
    #   hsclient
    #   pyzotero
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==1.4.0
    # via
    #   hs-restclient
    #   hsclient
requests-toolbelt==1.0.0
    # via hs-restclient
s3transfer==0.10.1
    # via boto3
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   furl
    #   html5lib
    #   isodate
    #   orderedmultidict
    #   python-dateutil
    #   rdflib
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
svglib==1.5.1
    # via easy-thumbnails
tinycss2==1.2.1
    # via
    #   cssselect2
    #   svglib
typing-extensions==4.10.0
    # via
    #   dj-database-url
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via
    #   botocore
    #   requests
uwsgi==2.0.24
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wheel==0.43.0
    # via pip-tools
whitenoise==6.6.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
